+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Change Log
 
-## [v5.3.0](https://github.com/plivo/plivo-dotnet/tree/v5.3.0) (2021-07-20)
-- This version includes advancements to the Messaging Interface that deals with the [Send SMS/MMS](https://www.plivo.com/docs/sms/api/message#send-a-message) interface, Creating a standard structure for `request/input` arguments to make implementation easier and incorporating support for the older interface.
+## [v5.3.0](https://github.com/plivo/plivo-dotnet/tree/v5.3.0) (2021-07-28)
+- Added generic error message for mandatory params in [Make Call](https://www.plivo.com/docs/voice/api/call#make-a-call) and [Multi party call](https://www.plivo.com/docs/voice/api/multiparty-call#start-a-new-multiparty-call).
+- Removed validation for `ringtimeout` and `delaydial` params in [Start a multi party call](https://www.plivo.com/docs/voice/api/multiparty-call#start-a-new-multiparty-call).
+
+## [v5.2.1](https://github.com/plivo/plivo-dotnet/tree/v5.2.1) (2021-07-27)
+- Updated default HTTP client request timeout to 5 seconds.
 
 ## [v5.2.0](https://github.com/plivo/plivo-dotnet/tree/v5.2.0) (2021-07-14)
 - Add support for MPC APIs and XML (Voice retry included), validated voice UTs.